Try to get around possible recursion issues with trap handling
It's possible that an ealier loop iteration is sneaking through on
multiple disarm attempts. To alleviate this problem, rather than
returning the result of the last invocation of `disarm_trap`, return the
value stored in `@trap_data[:is_disarmed]`, which is updated in process
and will be true if at _any_ point and iteration, the trap was disarmed
successfully.